If we try long enough and don't succeeded in renewing, give up the slave.
The way the code is written today, if the mansion gets rid of a lease, Jenkins appears to keep that slave around forever.
We keep trying to renew, which keeps failing.

I think it's fair to stop trying at some point, since we know that if we don't manage to renew a lease for too long.

This lets us clean up dead slaves