Dedicated NodeProvisioner for mansion slaves
This NodeProvisioner takes advantages of recent API introduced in Jenkins
to replace the default node provisioner used by Jenkins for Cloud plugins.
It allows to kick off node provisioning as soon as a job is in the build
queue. The patch is heavily based on Stephen's work on OC.

To limit the load on mansion server, no more than 2 nodes can be
provisioned at a time. Provisioning is considered complete when Jenkins
is able to start a build on the node.

This fixes the slow provisioning sometimes encountered and also
hopefully overprovisioning situations encountered when Mansion can't
provision nodes fast enough. In my recent testing, I was able to get 5
nodes for 1 build in the queue. With this patch, I never had more than
2 nodes provisioning at the same time.